alertdialog 추가

+++ lib/core/widgets/survey_card.dart
@@ -51,7 +51,76 @@ class _SurveyCardState extends State<SurveyCard> {
                   PopupMenuButton<String>(
                     onSelected: (value) {
                       if (value == 'delete') {
-                        widget.onDelete();
+                        showDialog<String>(
+                          context: context,
+                          builder: (BuildContext context) => AlertDialog(
+                            backgroundColor: Colors.white,
+                            shape: RoundedRectangleBorder(
+                              borderRadius: BorderRadius.circular(10)
+                            ),
+                            contentPadding: EdgeInsets.only(top: 36, bottom: 36),
+                            buttonPadding: EdgeInsets.all(0),
+                            insetPadding: EdgeInsets.all(58),
+                            iconPadding: EdgeInsets.all(0),
+                            titlePadding: EdgeInsets.all(0),
+                            actionsPadding: EdgeInsets.only(left: 16, right: 16, bottom: 13),
+                            content: Row(
+                              mainAxisAlignment: MainAxisAlignment.center,
+                              children: [
+                                const Text(
+                                  '해당 설문을 삭제하시겠습니까?',
+                                  style: TextStyle(
+                                    fontFamily: 'Pretendard',
+                                    fontWeight: FontWeight.w500,
+                                    fontSize: 17
+                                  ),
+                                ),
+                              ],
+                            ),
+                            actions: [
+                              Row(
+                                children: <Widget>[
+                                  TextButton(
+                                    style: ButtonStyle(
+                                      backgroundColor: WidgetStatePropertyAll(Color(0xffB4B4B4)),
+                                      shape: WidgetStatePropertyAll(RoundedRectangleBorder(
+                                        borderRadius: BorderRadius.circular(8)
+                                      )),
+                                      fixedSize: WidgetStatePropertyAll(Size(114, 39))
+                                    ),
+                                    onPressed: () => Navigator.pop(context, 'Cancel'),
+                                    child: const Text(
+                                      '아니요',
+                                      style: TextStyle(
+                                        color: Colors.white
+                                      ),
+                                    ),
+                                  ),
+                                  Expanded(child: Container()),
+                                  TextButton(
+                                    style: ButtonStyle(
+                                      backgroundColor: WidgetStatePropertyAll(Color(0xff164BC6)),
+                                      shape: WidgetStatePropertyAll(RoundedRectangleBorder(
+                                        borderRadius: BorderRadius.circular(8)
+                                      )),
+                                      fixedSize: WidgetStatePropertyAll(Size(114, 39))
+                                    ),
+                                    onPressed: () {
+                                      Navigator.pop(context, 'OK');
+                                      widget.onDelete();
+                                    },
+                                    child: const Text(
+                                      '네',
+                                      style: TextStyle(
+                                        color: Colors.white
+                                      ),
+                                    ),
+                                  ),
+                                ],
+                              ),
+                            ],
+                          ),
+                        );
                       }
                     },
                     itemBuilder: (context) => [
